// src/modules/hr/types/position.ts
export interface BaseDto {
  id: UUID;
  isDeleted: boolean;
  rowVersion: string;
  dateAdd: string;
  dateMod: string | null;
}

// Position Requirement Types
export interface PositionReqListDto extends BaseDto {
  professionType: ProfessionType;
  positionId: UUID;
  gender: '0' | '1' | '2';
  saturdayWorkOption: '0' | '1' | '2';
  sundayWorkOption: '0' | '1' | '2';
  workingHours: number;
  genderStr: string;
  saturdayWorkOptionStr: string;
  sundayWorkOptionStr: string;
  professionTypeStr: string;
}

export interface PositionReqAddDto {
  gender: '0' | '1' | '2';
  saturdayWorkOption: '0' | '1' | '2' | '3';
  sundayWorkOption: '0' | '1' | '2' | '3';
  workingHours: number;
  professionType: ProfessionType;
  positionId: UUID;
}

export interface PositionReqModDto {
  id: UUID;
  gender: '0' | '1' | '2';
  saturdayWorkOption: '0' | '1' | '2' | '3';
  sundayWorkOption: '0' | '1' | '2' | '3';
  workingHours: number;
  professionType: ProfessionType;
  positionId: UUID;
  rowVersion: string;
}